fix: GetSchemaByExpr() of other package
